feat: request create user and update user

+++ database/migrations/2023_07_15_032458_create_table_eletros.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('description');
             $table->string('tension');
             $table->string('brand');
-            $table->integer('price');
+            $table->decimal('price',8,2);
             $table->timestamps();
         });
     }
